Backend/requirements.txt
Flask
gunicorn
celery[redis]
redis
SQLAlchemy
psycopg2-binary
alembic
pdfplumber
python-docx
spacy
python-magic
python-dotenv
boto3
